fix(tweets): market.sim_day tweet, order yielding logic
The logic for guarding against Nones in the market.sim_day iterator was flawed and causing only values that had both a tweet and an order to get through, meaning the tables had the same counts. This corrects that and ensures all orders go through, and those that have tweets are only a portion.