chore(repo): Add lefthook (#115)
This commit adds [Lefthook](https://github.com/evilmartians/lefthook/)
as a way to manage Git Hooks and automatically run ruff, PyRight and a
commit message validator on commit.

For now, `ruff` and `pyright` validators are turned off since there are
many errors. They will be turned on once we have fixed the errors.